+++ packages/webawesome/docs/docs/resources/changelog.md
@@ -31,6 +31,9 @@ Web Awesome follows <a href="https://semver.org/" class="appearance-plain">Seman
 
 ## Unreleased
 
+:::added
+- Added the `<wa-random-content>` experimental component
+
 :::fixed
 
 - Fixed a bug in `<wa-chart>` and its variants that threw a `DataCloneError` when the Chart.js config contained functions, such as tooltip or scale callbacks